restructure the imports and file locations - move all files to parent folder which part of the bot - create package for tetue specific files
